+++ src/jsc/jsc_value.hpp
@@ -124,13 +124,18 @@ inline bool jsc::Value::is_valid(const JSValueRef &value) {
 }
 
 inline bool is_bson_type(JSContextRef ctx, const JSValueRef &value, std::string type) {
-    if (JSValueIsNull(ctx, proto) || JSValueIsUndefined(ctx, value) || !JSValueIsObject(ctx, value)) {
+    if (JSValueIsNull(ctx, value) || JSValueIsUndefined(ctx, value) || !JSValueIsObject(ctx, value)) {
         return false;
     }
 
 
     JSValueRef error = nullptr;
-    JSValueRef bsonType = JSObjectGetProperty(ctx, value, JSStringCreateWithUTF8CString("_bsontype"), &error);
+    JSObjectRef object = JSValueToObject(ctx, value, &error);
+    if (error) {
+        throw jsc::Exception(ctx, error);
+    }
+
+    JSValueRef bsonType = JSObjectGetProperty(ctx, object, JSStringCreateWithUTF8CString("_bsontype"), &error);
     if (error) {
         throw jsc::Exception(ctx, error);
     }
@@ -139,7 +144,13 @@ inline bool is_bson_type(JSContextRef ctx, const JSValueRef &value, std::string 
         return false;
     }
 
-    jsc::String bsonTypeValue = Value::to_string(ctx, bsonType);
+    jsc::String bsonTypeValue = JSValueToStringCopy(ctx, bsonType, &error);
+    // Since the string's retain value is +2 here, we need to manually release it before returning.
+    JSStringRelease(bsonTypeValue);
+    if (error) {
+        throw jsc::Exception(ctx, error);
+    }
+
     std::string bsonTypeStringValue = bsonTypeValue;
     return bsonTypeStringValue == type;
 }