+++ CHANGELOG.md
@@ -9,6 +9,8 @@ x.x.x Release notes (yyyy-MM-dd)
 * Fixed bugs so `Realm.User.auth.APIKeys` and `Realm.App.auth.EmailPassword` are properties and not functions.
 * Added missing methods to `Realm.Auth` classes.
 * When restarting an app, re-using the already logged in user would result in Sync not resuming.
+* Disabled client resync. (since v10.0.0-beta.1)
+* Android report version as number and not as string, preventing an app to launch. (since v10.0.0-beta.1)
 
 ### Compatibility
 * MongoDB Realm Cloud.